Extract poiGroup
